add ScenarioManagerPDOTest
