+++ Tests/bootstrap.php
@@ -1,4 +1,3 @@
 <?php
-const TEST_DB_PATH = __DIR__ . "\\database\\activhome-test.db";
-const TEST_INIT_SQL_PATH = __DIR__ . "\\sql\\Scenarios.sql";
-const SQLITE_PATH="C:\\sqlite\\sqlite3.exe";
+const TEST_DB_PATH = __DIR__ . "/database/activhome-test.db";
+const TEST_INIT_SQL_PATH = __DIR__ . "/sql/Scenarios.sql";
